% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_territories.R
\name{estimate_territories}
\alias{estimate_territories}
\title{Estimate the number of territories in a catchment}
\usage{
estimate_territories(
  forage_raster,
  confirm_signs,
  low_thresh = 0,
  upper_thresh = 0.95
)
}
\description{
Uses the kernel density maps to estimate the area of territories by locating areas of activity that
contain foraging intensity above a given threshold, which locates potential territories - to confirm
territories, confirmatory signs can be provided including dam and dwelling locations.
}
