% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_territories.R
\name{gen_territories}
\alias{gen_territories}
\title{Caller function for create territories wrapped in safely.}
\usage{
gen_territories(BeaverNetwork, progbar = TRUE, multicore = TRUE, ncores)
}
\description{
Caller function for create territories wrapped in safely.
}
