% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_map_features.R
\name{get_rivers}
\alias{get_rivers}
\title{Retrieve an Open Street Map River Network for a given catchment area}
\usage{
get_rivers(catchment)
}
\arguments{
\item{catchment}{An sf object or an sf-readable file of the catchment area or area of interest.
See sf::st_drivers() for available drivers.}
}
\description{
This function allows you to download a river network layer in 'sf' format for
a catchment area of your choice. This can be useful when making many plots of the
same area and saves frequent calls to the OSM server for each plot when using river = TRUE
in the \code{beavertools::terr_plot()} and \code{beavertools::plot_forage_density()} functions.
}
\examples{
# this generates an sf object for the river network of the R. Otter catchment.
rivs <- get_rivers(RivOtter_Catch_Area)
plot(sf::st_geometry(rivs))
}
