% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflate_bounds.R
\name{inflate_bbox}
\alias{inflate_bbox}
\title{A function to provide an inflated bbox for input to Kernel Density function}
\usage{
inflate_bbox(sp_obj, value)
}
\description{
A function to provide an inflated bbox for input to Kernel Density function
}
