% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_plot.R
\name{panel_plot}
\alias{panel_plot}
\title{generate a panel plot of multiple maps generated with \code{plot_territories} or \code{plot_forage_density}}
\usage{
panel_plot(
  terr_plot_list,
  scalebar = TRUE,
  scalebar_loc = "tl",
  north_arrow = TRUE,
  north_arrow_loc = "br",
  north_arrow_size = 0.5
)
}
\description{
generate a panel plot of multiple maps generated with \code{plot_territories} or \code{plot_forage_density}
}
