% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde_plot.R
\name{panel_plot_forage_density}
\alias{panel_plot_forage_density}
\title{generate a panel plot of multiple forage density plots generated with \code{plot_forage_density}}
\usage{
panel_plot_forage_density(
  forage_plot_list,
  scalebar = TRUE,
  scalebar_loc = "tl",
  north_arrow = TRUE,
  north_arrow_loc = "br",
  north_arrow_size = 0.5
)
}
\description{
generate a panel plot of multiple forage density plots generated with \code{plot_forage_density}
}
