% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RivOtter_Catch_Area}
\alias{RivOtter_Catch_Area}
\title{River Otter Catchment Area.}
\format{
An sf object with 1 rows and 3 variables:
\describe{
\item{Name}{Simply the name of the catchment}
\item{geometry}{The geometry data for the catchment}
}
}
\source{
generated from OS Terrain 5: \url{https://www.ordnancesurvey.co.uk/business-government/products/terrain-5}
using GRASS GIS \url{https://grass.osgeo.org/}
}
\usage{
RivOtter_Catch_Area
}
\description{
A dataset ('sf' object) provides the River Otter Catchment area. This can be used in various plotting
functions to mask areas outside of the Otter catchment.
}
\keyword{datasets}
