% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RivOtter_OtherSigns}
\alias{RivOtter_OtherSigns}
\title{Beaver Dwelling and Dam Locations recorded during the River Otter Beaver Trial.}
\format{
An sf object with 3683 rows and 4 variables:
\describe{
\item{SurveySeason}{The winter season during which the sign was surveyed}
\item{RecordDate}{The date that a sign was recorded}
\item{FeedCat}{Character String describing the effort exerted by beaver on a given sign}
\item{geometry}{The geometry data for the point location}
}
}
\source{
\url{https://www.exeter.ac.uk/creww/research/beavertrial/}
}
\usage{
RivOtter_OtherSigns
}
\description{
A dataset ('sf' object) which includes the locations, survey times and
survey season for beaver dwellings and dams. These data can be used to confirm or
reject the automated territory status. Essentially, it is considered that where dams
or dwellings occur, an area of activity can be considered as a territory.
}
\keyword{datasets}
