% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_terr_cap.R
\name{sim_terr_cap}
\alias{sim_terr_cap}
\title{A function to run a territory capacity simulation}
\usage{
sim_terr_cap(
  BeaverNetwork,
  n_p_terr_sim = 1,
  n_hab_sim = 5,
  min_veg = c(2.5, 4),
  progbar = TRUE
)
}
\description{
This method enables for the calculation of uncertainty based on n random territory simulations and a specified
range of minimum habitat requirement.
}
