% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_territories.R
\name{user_classify}
\alias{user_classify}
\title{Add user specified classification of territory areas}
\usage{
user_classify(territory_poly, territory = NULL, possible = NULL)
}
\description{
Opportunity to add a new column 'user_class' which enables the user to confirm or reject
the automated assessment of territory areas.
}
