% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_maps.R
\name{animate_maps}
\alias{animate_maps}
\title{Function to Animate a list of maps generated with either \code{plot_territories} or \code{plot_forage_density}}
\usage{
animate_maps(
  map_list,
  filename = NULL,
  x_pix = 1200,
  y_pix = 675,
  pix_res = 96,
  n_frames = 10,
  fps = 5
)
}
\description{
Function to Animate a list of maps generated with either \code{plot_territories} or \code{plot_forage_density}
}
