% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fkd.R
\name{forage_density}
\alias{forage_density}
\title{Function to generate Kernel Density plot from Beaver Forage Data}
\usage{
forage_density(
  forage_points,
  impact_cat,
  grid_size = 20,
  kern_bw = 250,
  kd_extent,
  kd_weights = c(1, 1000, 1e+06),
  low_thresh = 1e-12,
  standardise = FALSE
)
}
\description{
This function allows users to generate a kernel density raster from a collection of points.
}
