% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopCarryCap.R
\name{gen_territories}
\alias{gen_territories}
\title{Caller function for create territories wrapped in safely.}
\usage{
gen_territories(BeaverNetwork)
}
\description{
Caller function for create territories wrapped in safely.
}
