% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/territory_cap.R
\name{territory_cap}
\alias{territory_cap}
\title{Beaver territory capacity simulation}
\usage{
territory_cap(
  territories,
  min_veg = 2.5,
  min_bdc = 1,
  progbars = TRUE,
  multicore = FALSE,
  ncores
)
}
\description{
Beaver territory capacity simulation
}
